Add __telemetry_version__ to event capsule
We might add more fields or change the meaning of fields in
the capsule itself. This lets us do so in a way that is backwards
compatible and explicit. Analysts can use this field to determine
exactly what data they are working with.

We rename __version__ to __schema_version__ to distinguish
this now. This is a backwards incompatible change, but it's
early enough to be ok